requirements.txt
fastapi==0.115.0
uvicorn==0.30.1
python-binance==1.0.19
requests==2.32.3
python-dotenv==1.0.1
pandas==2.2.3
numpy==2.1.3
